New Model_Parameters.txt file
Updated Model_Parameters.txt with the following changes:
1) Removed the parameter “Iterate Over” and also the “Fixed body” parameters.  I have a simpler and more flexible idea to set up the program (which Andrew also liked), namely to set up the nested for-loops to automatically iterate over both body size and body number.  If we want to iterate over only one of those parameters, we can simply set the starting and ending values of the other parameter to be the same.  
2)	I added parameters for the step size of mu and sigma to use in the iteration (for both body size and body number).  That way this can also be user controlled.
I presume the iterations will be implemented using range functions followed by some for loops.  Andrew pointed out that we’ll need to remember to make the ending value of the range function one step larger than the ending mu or sigma, since the range function is non-inclusive of the ending value.